+++ constant/build.go
@@ -4,6 +4,6 @@ const Version = "3.8.0"
 
 var (
 	BuiltAt   string
-	BuiltBy   string
-	GitCommit string
+	BuiltBy  string
+	Revision string
 )