+++ updater/update.go
@@ -12,29 +12,29 @@ import (
 
 // Update updates mangal to the latest version.
 func Update() (err error) {
-	method := detectInstallationMethod()
+	method := DetectInstallationMethod()
 
 	switch method {
-	case homebrew:
+	case Homebrew:
 		fmt.Printf("%s Homebrew installation detected", icon.Get(icon.Progress))
 		return updateHomebrew()
-	case scoop:
+	case Scoop:
 		fmt.Printf("%s Scoop installation detected", icon.Get(icon.Progress))
 		return updateScoop()
-	case termux:
+	case Termux:
 		fmt.Printf("%s Termux installation detected", icon.Get(icon.Progress))
 		return updateScript()
-	case script:
+	case Script:
 		fmt.Printf("%s Script installation detected", icon.Get(icon.Progress))
 		return updateScript()
-	case unknown:
-		return errors.New("unknown installation method, can't update")
+	case Unknown:
+		return errors.New("Unknown installation method, can't update")
 	}
 
 	return
 }
 
-// updateHomebrew updates mangal using homebrew.
+// updateHomebrew updates mangal using Homebrew.
 func updateHomebrew() (err error) {
 	cmd := exec.Command("brew", "upgrade", constant.Mangal)
 	cmd.Stdout = os.Stdout
@@ -42,15 +42,15 @@ func updateHomebrew() (err error) {
 	return cmd.Run()
 }
 
-// updateScoop updates mangal using scoop.
+// updateScoop updates mangal using Scoop.
 func updateScoop() (err error) {
-	cmd := exec.Command("scoop", "update", constant.Mangal)
+	cmd := exec.Command("Scoop", "update", constant.Mangal)
 	cmd.Stdout = os.Stdout
 	cmd.Stderr = os.Stderr
 	return cmd.Run()
 }
 
-// updateScript updates mangal using the script.
+// updateScript updates mangal using the Script.
 func updateScript() (err error) {
 	res, err := http.Get(constant.InstallScriptURL)
 	if err != nil {
@@ -58,7 +58,7 @@ func updateScript() (err error) {
 	}
 
 	if res.StatusCode != http.StatusOK {
-		return fmt.Errorf("error fetching script: status code %d", res.StatusCode)
+		return fmt.Errorf("error fetching Script: status code %d", res.StatusCode)
 	}
 
 	var scriptSource []byte