+++ CHANGELOG.md
@@ -18,8 +18,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
   `downloader.download_cover` (default: `false`)
 - Better progress message while downloading in TUI mode
 - Set option `downloader.create_volume_dir` to `false` by default
-- Version command now show more information (such as build date, commit hash, etc.)
+- Version command now shows more information (such as build date, commit hash, etc.)
 - New flag for inline mode: `--output/-o` to redirect output to file (will use STDOUT if not set)
+- New `mangal config set` command to set config values. See `mangal help config set` for more info.
 
 ## 3.7.0